+++ authlib/little_auth/admin.py
@@ -1,15 +1,49 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as StockUserAdmin
+from django.utils.translation import gettext_lazy as _
 
 from authlib.little_auth.models import User
 
 
 @admin.register(User)
 class UserAdmin(StockUserAdmin):
-    fieldsets = None
     add_fieldsets = (
-        (None, {"classes": ("wide",), "fields": ("email", "password1", "password2")}),
+        (None, {"classes": ["wide"], "fields": ("email", "password1", "password2")}),
     )
+    fieldsets = [
+        (
+            None,
+            {
+                "fields": [
+                    "is_active",
+                    "email",
+                    "password",
+                    "full_name",
+                ]
+            },
+        ),
+        (
+            _("Permissions"),
+            {
+                "fields": [
+                    "is_staff",
+                    "is_superuser",
+                    "role",
+                ]
+            },
+        ),
+        (
+            _("Advanced"),
+            {
+                "classes": ["collapse"],
+                "fields": [
+                    "date_joined",
+                    "last_login",
+                    "groups",
+                ],
+            },
+        ),
+    ]
     list_display = (
         "email",
         "full_name",
@@ -24,3 +58,4 @@ class UserAdmin(StockUserAdmin):
     search_fields = ("full_name", "email")
     filter_horizontal = ("groups", "user_permissions")
     radio_fields = {"role": admin.VERTICAL}
+    readonly_fields = ["last_login"]