+++ authlib/little_auth/migrations/0001_initial.py
@@ -49,7 +49,11 @@ class Migration(migrations.Migration):
                 ),
                 (
                     "is_staff",
-                    models.BooleanField(default=False, verbose_name="is staff"),
+                    models.BooleanField(
+                        default=False,
+                        help_text="Designates whether the user can log into this admin site.",
+                        verbose_name="is staff",
+                    ),
                 ),
                 (
                     "date_joined",